refactor: (#152) 신고 컴포넌트에 사용되는 문자열 상수화 스크롤 이동 유틸 함수 파일 경로 이동, CSS 속성 변경
