+++ frontend/src/components/report/CommentReportModal/constants.ts
@@ -1,6 +1,13 @@
+export const COMMENT_REPORT_KIND = {
+  BEHAVIOR: 'behavior',
+  SPAMMING: 'spamming',
+  ADVERTISING: 'advertising',
+  EXPLICIT: 'explicit',
+} as const;
+
 export const COMMENT_REPORT_MESSAGE = {
-  behavior: '부적절한 언행',
-  spamming: '도배성 댓글 작성',
-  advertising: '광고성 댓글 작성',
-  explicitContent: '성적인 댓글 작성',
+  [COMMENT_REPORT_KIND.BEHAVIOR]: '부적절한 언행',
+  [COMMENT_REPORT_KIND.SPAMMING]: '도배성 댓글 작성',
+  [COMMENT_REPORT_KIND.ADVERTISING]: '광고성 댓글 작성',
+  [COMMENT_REPORT_KIND.EXPLICIT]: '성적인 댓글 작성',
 };
